scala-js: enable JSPlatform
