+++ shared/src/main/scala/scorex/crypto/authds/merkle/sparse/SparseMerkleTree.scala
@@ -1,6 +1,5 @@
 package scorex.crypto.authds.merkle.sparse
 
-import scorex.utils.Longs
 import scorex.crypto.authds.LeafData
 import scorex.crypto.hash._
 
@@ -174,139 +173,5 @@ case class SparseMerkleProof[D <: Digest](idx: Node.ID,
   def valid(tree: SparseMerkleTree[D])(implicit hf: CryptographicHash[D]): Boolean = valid(tree.rootDigest, tree.height)
 }
 
-object OpsBenchmark extends App {
-  implicit val hf: CryptographicHash[Digest32] = new Blake2b256Unsafe
-
-  val height: Byte = 50
-
-  var tree = SparseMerkleTree.emptyTree(height)
-
-  //bootstrapping
-
-  val bSize = 10000
-
-  var proof: SparseMerkleProof[Digest32] = null
-
-  val tb0 = System.currentTimeMillis()
-  (0 until bSize).foreach {i =>
-    if(i == bSize/2) proof = tree.lastProof
-
-    val proofsToUpdate = if(i >= bSize/2) {
-      Seq(proof)
-    } else Seq.empty[SparseMerkleProof[Digest32]]
-
-    if(i%1000 == 0) println(s"$i elements added")
-    val (t, p) = tree.update(tree.lastProof, Some(LeafData @@ Longs.toByteArray(i)), proofsToUpdate).get
-
-    tree = t
-    if(i >= bSize/2) proof = p.head
-  }
-  val tb = System.currentTimeMillis()
-  println(s"bootstrapping time: ${tb-tb0}")
-
-  val tv0 = System.currentTimeMillis()
-  (1 to 1000).foreach(i => proof.valid(tree))
-  val tv = System.currentTimeMillis()
-  println(s"1000 updates time: ${tv-tv0} ms")
 
 
-
-    val te0 = System.currentTimeMillis()
-    (1 to 1000).foreach(_ =>
-      tree.update(tree.lastProof, Some(LeafData @@ Longs.toByteArray(2)), Seq())
-    )
-    val te = System.currentTimeMillis()
-
-    val tp0 = System.currentTimeMillis()
-    (1 to 1000).foreach(_ =>
-      tree.update(tree.lastProof, Some(LeafData @@ Longs.toByteArray(2)), Seq(proof))
-    )
-    val tp = System.currentTimeMillis()
-
-    val tu = (tp-tp0) - (te-te0)
-
-    println("1000 proof updates: " + tu)
-}
-
-object BlockchainSimulator extends App {
-
-  type PubKey = Array[Byte]
-
-  val PubKeyLength = 32
-
-  implicit val hf: CryptographicHash[Digest32] = new Blake2b256Unsafe
-
-  case class Transaction(amount: Long,
-                         sender: PubKey,
-                         recipient: PubKey,
-                         coinBalance: Long,
-                         coinProof: SparseMerkleProof[Digest32])
-
-  object Transaction {
-    def coinBytes(pubKey: PubKey, balance: Long) = Some(LeafData @@ (pubKey ++ Longs.toByteArray(balance)))
-
-    def process(tx: Transaction,
-                state: SparseMerkleTree[Digest32]):
-    Try[(SparseMerkleTree[Digest32], Seq[SparseMerkleProof[Digest32]])] = Try {
-
-      require(tx.amount <= tx.coinBalance)
-      require(tx.coinProof.leafDataOpt.get sameElements coinBytes(tx.sender, tx.coinBalance).get)
-      require(tx.coinProof.valid(state.rootDigest, height))
-
-      val (state1, _) = state.update(tx.coinProof, None).get
-
-      val (state2, proofs2) = state1.update(state1.lastProof,
-        coinBytes(tx.recipient, tx.amount),
-        Seq(state1.lastProof)).get
-
-      if (tx.amount == tx.coinBalance) state2 -> proofs2
-      else state2.update(state2.lastProof,
-        coinBytes(tx.sender, tx.coinBalance - tx.amount),
-        proofs2 :+ state2.lastProof).get
-    }
-  }
-
-  case class Block(transactions: Seq[Transaction])
-
-  val txsCache = new mutable.ArrayBuffer()
-  val maxTxsCacheSize = 5000
-
-  val txsPerBlock = 1000
-  val numOfBlocks = 1000000
-
-  val height = 30: Byte
-
-  val godAccount = Array.fill(32)(0: Byte)
-  val godBalance = 100000000000L //100B
-
-  val emptyState = SparseMerkleTree.emptyTree(height)
-  val (initialState, godProofs) = emptyState.update(emptyState.lastProof,
-    Transaction.coinBytes(godAccount, godBalance),
-    Seq(emptyState.lastProof)).get
-
-  var godProof = godProofs.head
-  var currentGodBalance = godBalance
-  val txAmount = 10
-
-  (1 to numOfBlocks).foldLeft(initialState) { case (beforeBlocktree, blockNum) =>
-    val (afterTree, processingTime) = (1 to txsPerBlock).foldLeft(beforeBlocktree -> 0L) { case ((tree, totalTime), txNum) =>
-      val recipient = hf(scala.util.Random.nextString(20))
-
-      val tx = Transaction(txAmount, godAccount, recipient, currentGodBalance, godProof)
-
-      val t0 = System.currentTimeMillis()
-      val (updState, proofs) = Transaction.process(tx, tree).get //we generate always valid transaction
-      val t = System.currentTimeMillis()
-
-      currentGodBalance = currentGodBalance - txAmount
-      godProof = proofs.last
-
-      updState -> (totalTime + (t - t0))
-    }
-
-    println(s"Block $blockNum, processing time: $processingTime ms")
-    println(godProof)
-
-    afterTree
-  }
-}
